Moved Features component to own folder
