+++ src/App.tsx
@@ -1,7 +1,7 @@
 import Header from './components/organisms/Header/Header.tsx';
 import Footer from './components/organisms/Footer/Footer.tsx';
 import Hero from './components/organisms/Hero/Hero.tsx';
-import Features from './components/organisms/Features.tsx';
+import Features from './components/organisms/Features/Features.tsx';
 import Downloads from './components/organisms/Downloads/Downloads.tsx';
 import Faqs from './components/organisms/Faqs/Faqs.tsx';
 import Contact from './components/organisms/Contact/Contact.tsx';
